pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "polymarket-arb"
version = "0.1.0"
description = "Read-only Polymarket analytics scanner for dislocation research."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Codex" },
]
dependencies = [
  "fastapi>=0.115.0,<0.116.0",
  "httpx>=0.27.0,<0.29.0",
  "pydantic>=2.8.0,<3.0.0",
  "pydantic-settings>=2.3.0,<3.0.0",
  "PyYAML>=6.0.0,<7.0.0",
  "typer>=0.18.0,<0.19.0",
  "uvicorn>=0.30.0,<0.31.0",
  "websockets>=13.1,<14.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.10.0,<2.0.0",
  "pytest>=8.2.0,<9.0.0",
  "pytest-asyncio>=0.23.0,<0.24.0",
  "ruff>=0.5.0,<0.6.0",
]

[project.scripts]
polymarket-arb = "polymarket_arb.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/polymarket_arb"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "B",
  "UP",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
mypy_path = "src"
packages = ["polymarket_arb"]
